+++ processor/src/test/resources/input/com/example/inject/FactoryConsumer2Model.java
@@ -1,0 +1,31 @@
+package com.example.inject;
+
+import arez.annotations.Action;
+import arez.annotations.ArezComponent;
+import arez.annotations.InjectMode;
+import arez.annotations.PerInstance;
+import arez.annotations.PostConstruct;
+import javax.inject.Inject;
+import javax.inject.Singleton;
+
+@Singleton
+@ArezComponent( inject = InjectMode.CONSUME )
+public abstract class FactoryConsumer2Model
+{
+  @Inject
+  Runnable _action;
+
+  FactoryConsumer2Model( @PerInstance final int count, final long nonPerInstanceValue )
+  {
+  }
+
+  @PostConstruct
+  final void postConstruct()
+  {
+  }
+
+  @Action
+  public void myActionStuff()
+  {
+  }
+}